requirements.txt
# Core framework
streamlit>=1.31.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine learning / NLP
scikit-learn>=1.3.0

# API requests
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Optional: For advanced visualizations (Phase 3)
plotly>=5.18.0
matplotlib>=3.7.0

# Optional: For testing
pytest>=7.4.0
